BE-flask-fix/requirements.txt
waitress
cryptography
requests
flask
Flask-CORS
langchain
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
huggingface-hub
chromadb
pydantic
torch
transformers
accelerate
numpy
protobuf
gunicorn
marker-pdf[full]
pypdf2
unstructured[md]
